// packages/core/src/types/enum/date-system.ts
/**
 * Copyright 2023-present DreamNum Co., Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/** Workbook date serial systems supported by Univer. */
export enum DateSystem {
    /** Serial 1 is 1900-01-01; serials 0 and 60 preserve Excel's compatibility-only dates. */
    Date1900 = 'date1900',
    /** Serial 0 is 1904-01-01. */
    Date1904 = 'date1904',
}